SELinux: drop several type enforcement rules, so that a module could install on Fedora 33
I don't really understand what's going on, but it's apparent that there is no such type as `shell_t` on Fedora 33.
It could be removed at some point, but I have no idea how to investigate this.
As for me, I periodically use `audit2why`/`audit2allow`, so this is mostly harmless.
Yet failure in `prltimesync` module causes a failure in `prltoolsd` module, so this is sort of not okay.